+++ frontend/src/components/Header.tsx
@@ -5,7 +5,7 @@ export const Header = () => {
   return (
     <header className="flex items-center justify-between px-6 py-4 border-b border-border/50">
       <div>
-        <img src="src/assets/logo.png" alt="Logo da empresa" className="w-28" />
+        <img src="src/assets/logo.png" alt="Logo da empresa" className="sm:w-28 w-24" />
       </div>
 
       <div className="flex items-center gap-2">